Revert "Salutation and Gender in Contact (#3079)"
This reverts commit f52e389d4f53ee1a6015071681532801dd54cc94.
