--- frappe/contacts/doctype/gender/gender.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2017, Frappe Technologies and contributors
-// For license information, please see license.txt
-
-frappe.ui.form.on('Gender', {
-	refresh: function() {
-
-	}
-});
