--- frappe/contacts/doctype/gender/gender.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2017, Frappe Technologies and contributors
-# For license information, please see license.txt
-
-from __future__ import unicode_literals
-from frappe.model.document import Document
-
-class Gender(Document):
-	pass
