--- frappe/contacts/doctype/gender/test_gender.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2017, Frappe Technologies and Contributors
-# See license.txt
-from __future__ import unicode_literals
-
-import unittest
-
-class TestGender(unittest.TestCase):
-	pass
