--- frappe/contacts/doctype/salutation/test_salutation.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2017, Frappe Technologies and Contributors
-# See license.txt
-from __future__ import unicode_literals
-
-import unittest
-
-class TestSalutation(unittest.TestCase):
-	pass
