+++ frappe/geo/doctype/address/address.js
@@ -15,7 +15,7 @@ frappe.ui.form.on("Address", {
 		}
 		frm.set_query('link_doctype', "links", function() {
 			return {
-				query: "frappe.contacts.address_and_contact.filter_dynamic_link_doctypes",
+				query: "frappe.geo.address_and_contact.filter_dynamic_link_doctypes",
 				filters: {
 					fieldtype: "HTML",
 					fieldname: "address_html",
