+++ frappe/geo/doctype/address/test_records.json
@@ -1,0 +1,13 @@
+[
+ {
+  "address_line1": "_Test Address Line 1",
+  "address_title": "_Test Address",
+  "address_type": "Office",
+  "city": "_Test City",
+  "state": "Test State",
+  "country": "India",
+  "doctype": "Address",
+  "is_primary_address": 1,
+  "phone": "+91 0000000000"
+ }
+]
