+++ frappe/geo/doctype/address_template/address_template.json
@@ -1,6 +1,5 @@
 {
  "allow_copy": 0, 
- "allow_guest_to_view": 0, 
  "allow_import": 0, 
  "allow_rename": 1, 
  "autoname": "field:country", 
@@ -24,7 +23,6 @@
    "ignore_user_permissions": 0, 
    "ignore_xss_filter": 0, 
    "in_filter": 0, 
-   "in_global_search": 0, 
    "in_list_view": 1, 
    "in_standard_filter": 1, 
    "label": "Country", 
@@ -54,7 +52,6 @@
    "ignore_user_permissions": 0, 
    "ignore_xss_filter": 0, 
    "in_filter": 0, 
-   "in_global_search": 0, 
    "in_list_view": 1, 
    "in_standard_filter": 0, 
    "label": "Is Default", 
@@ -84,7 +81,6 @@
    "ignore_user_permissions": 0, 
    "ignore_xss_filter": 0, 
    "in_filter": 0, 
-   "in_global_search": 0, 
    "in_list_view": 0, 
    "in_standard_filter": 0, 
    "label": "Template", 
@@ -102,20 +98,20 @@
    "unique": 0
   }
  ], 
- "has_web_view": 0, 
  "hide_heading": 0, 
  "hide_toolbar": 0, 
  "icon": "fa fa-map-marker", 
  "idx": 0, 
  "image_view": 0, 
  "in_create": 0, 
+ "in_dialog": 0, 
  "is_submittable": 0, 
  "issingle": 0, 
  "istable": 0, 
  "max_attachments": 0, 
- "modified": "2017-04-10 13:09:53.761009", 
+ "modified": "2017-01-13 05:11:37.499528", 
  "modified_by": "Administrator", 
- "module": "Contacts", 
+ "module": "Geo", 
  "name": "Address Template", 
  "name_case": "", 
  "owner": "Administrator", 
@@ -130,6 +126,7 @@
    "export": 1, 
    "if_owner": 0, 
    "import": 0, 
+   "is_custom": 0, 
    "permlevel": 0, 
    "print": 0, 
    "read": 1, 
@@ -144,7 +141,6 @@
  "quick_entry": 1, 
  "read_only": 0, 
  "read_only_onload": 0, 
- "show_name_in_global_search": 0, 
  "sort_field": "modified", 
  "sort_order": "DESC", 
  "track_changes": 0,
