+++ frappe/geo/report/addresses_and_contacts/addresses_and_contacts.json
@@ -7,9 +7,9 @@
  "doctype": "Report", 
  "idx": 2, 
  "is_standard": "Yes", 
- "modified": "2017-04-10 15:04:12.498920", 
+ "modified": "2017-02-24 19:57:37.368498", 
  "modified_by": "Administrator", 
- "module": "Contacts", 
+ "module": "Geo", 
  "name": "Addresses And Contacts", 
  "owner": "Administrator", 
  "ref_doctype": "Address",
