+++ frappe/geo/report/addresses_and_contacts/addresses_and_contacts.py
